Zsh: Moved vi-bindings to zshrc
Removed old vim-mode plugin, which was outdated and bloated. Zsh has
since fixed a number of the problems that plugin was trying to solve
when I wrote it, such as multi-level undos. Additionally, zsh has added
visual mode, which the old plugin was interfering with.

The plugin as now been replaced with a handful of lines in zshrc.